agent/pyproject.toml
[tool.poetry]
name = "clinical-trial-agent"
version = "1.0.0"
description = "LangGraph-based AI agent for clinical trial site selection"
authors = ["Clinical Trial Systems Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
langchain-anthropic = "^0.2.0"
anthropic = "^0.39.0"
httpx = "^0.27.0"
pydantic = "^2.9.0"
python-dotenv = "^1.0.0"
typing-extensions = "^4.12.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
